requirements.txt
fastapi
uvicorn
openai
tortoise-orm
passlib
bcrypt
python-multipart
PyJWT
sqlalchemy
bcrypt
gaurdrails-ai
dotenv
os

